+++ Xcode/ci_scripts/ci_post_clone.sh
@@ -6,9 +6,11 @@ git clone --depth 1 --branch "v1.4.1" "https://github.com/nodenv/nodenv.git" ~/.
 export PATH="$HOME/.nodenv/bin:$PATH"
 
 mkdir -p "$(nodenv root)"/plugins
-git clone --depth 1 --branch "v4.9.138" "https://github.com/nodenv/node-build.git" "$(nodenv root)"/plugins/node-build
+git clone --depth 1 "https://github.com/nodenv/node-build.git" "$(nodenv root)"/plugins/node-build
 
 pushd ..
 nodenv install
-nodenv exec npm ci --no-fund
+nodenv exec corepack enable pnpm
+nodenv rehash
+nodenv exec pnpm i
 popd