+++ src/background/index.js
@@ -1,9 +1,9 @@
-import browser from 'webextension-polyfill';
+import browser from '../common/browser.js';
 
 import _catch from '../common/catch.js';
 import download from '../common/download.js';
 
-browser.runtime.onMessage.addListener(_catch((message, sender, sendResponse) => {
+browser.runtime.onMessage.addListener(_catch((message, sender) => {
     console.log("Received request: ", message, sender);
 
     const [key, value] = Object.entries(message)[0];
@@ -13,16 +13,11 @@ browser.runtime.onMessage.addListener(_catch((message, sender, sendResponse) => 
             const downloads = value.downloads;
             const originURL = value.originURL;
             
-            download(downloads, originURL)
-                .then(() => {
-                    sendResponse();
-                })
-                .catch((reason) => {
-                    console.error(reason);
-                    sendResponse(new Error(reason));
+            return download(downloads, originURL)
+                .catch((e) => {
+                    console.error(e);
+                    throw e
                 });
-
-            return true;
         default:
             return false;
     }