+++ src/popup/popup.js
@@ -1,9 +1,7 @@
-import browser from 'webextension-polyfill';
-
 import _catch from '../common/catch.js';
+import { default as browser, fetchCurrentTab } from '../common/browser.js';
 
 import './popup.css';
-import { prepareDownloadForBackground } from '../common/download.js';
 
 /**
  * 
@@ -19,55 +17,47 @@ const setDisabled = (element, disabled) => {
 }
 
 async function downloadAll(media, originURL) {
-    const downloads = media.map((media) => {
-        return {
-            filename: media.download,
-            url: media.href
-        }
-    });
-
-    const preparedDownloads = await prepareDownloadForBackground(downloads, originURL);
-
     await browser.runtime.sendMessage({
-        'download': preparedDownloads
+        'download': {
+            'downloads': media,
+            originURL
+        }
     });
 }
 
-async function updateBody(media, senderURL) {
+async function updateBody(_media, senderURL) {
     const downloadAllButton = document.getElementById('download-all-button');
 
-    const _media = media ?? [];
+    const media = _media ?? [];
 
-    setDisabled(downloadAllButton, !(_media.length > 0));
+    setDisabled(downloadAllButton, !(media.length > 0));
     downloadAllButton.onclick = _catch(async () => {
         const disabled = downloadAllButton.hasAttribute('disabled');
 
         setDisabled(downloadAllButton, true);
 
-        await downloadAll(_media, senderURL);
+        await downloadAll(media, senderURL);
 
         setDisabled(downloadAllButton, disabled);
     });
 
     const donwloadAllButtonDescription = downloadAllButton.getElementsByClassName('description')[0];
-    donwloadAllButtonDescription.textContent = `Total ${_media.length} ${_media.length == 1 ? 'file' : 'files'}`;
+    donwloadAllButtonDescription.textContent = `Total ${media.length} ${media.length == 1 ? 'file' : 'files'}`;
 }
 
 _catch(async () => {
     updateBody();
 
-    const currentTab = (await browser.tabs.query({
-        active: true,
-        currentWindow: true
-    }))[0];
+    const currentTab = await fetchCurrentTab();
 
     let fetchPort = browser.tabs.connect(currentTab.id, { name: 'fetch' });
+    fetchPort.onMessage.addListener((message) => {
+        console.log("Received request: ", message);
+        updateBody(message.media, currentTab.url);
+    });
     fetchPort.onDisconnect.addListener((port) => {
         if (port.error) {
             console.error(port.error);
         }
-    });
-    fetchPort.onMessage.addListener((message) => {
-        updateBody(message.media, currentTab.url);
     });
 })();